[#2652] feat(spark): Introduce partition records number check to ensure data consistency (#2653)
### What changes were proposed in this pull request?

This PR ensures end-to-end data consistency by verifying the record counts of each partition.
In this initial step, ShuffleWriteTaskStats is introduced to store record counts for validation.
In the next phase, this mechanism will be extended to support row-level checksums and block count verification.

I have only validated this patch on Spark 3.5.0, and this feature is enabled only for Spark versions at least `3.5.0`

### Why are the changes needed?

for the #2652

### Does this PR introduce _any_ user-facing change?

`spark.rss.client.integrityValidation.enabled=false` . 

### How was this patch tested?

Unit tests.