[#2665] feat(spark): Reconstruct the shuffle handle from initial spark handle it haven't bee updated (#2667)
### What changes were proposed in this pull request?

Reconstruct the shuffle handle information from static spark handle if it hasn’t been updated. This means the latest shuffle handle data will no longer need to be transmitted through the RPC layer, reducing the load on the driver—especially when there are a large number of tasks.

### Why are the changes needed?

For issue #2665: Once partition reassignment is enabled, the shuffle information is always retrieved when a task starts, which puts significant pressure on the Spark driver. Although in most cases the shuffle information remains unchanged, this behavior provides a natural point for optimization.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing tests could cover this PR's change.