[#2673] feat(spark)(part-1): Add client-side support for storing partition stats on shuffle servers (#2669)
### What changes were proposed in this pull request?

This is the part-1 PR only with uniffle client changes of making the partition stats stored in the shuffle-server side to make the integrity validation mechanism more stable. BTW, the shuffle-servers side changes will be implemented in the further PRs, and this PR is also compatible with  the legacy shuffle-server protocol.

### Why are the changes needed?

the subtask for the issue #2673.

By leveraging the PR #2653 , we could end-to-end ensure the data consistency. But, the partition stats stored in the spark driver side, for the normal spark stages, this design runs well. But with the 100000 tasks with 10000 partitions, this will make the Spark driver overload. From the point of cluster spark jobs, some huge jobs will hang when getting the blockManagerIds, that will cost almost 20mins for one reader task, that is unacceptable. 

And so, this PR implements the server side store the partition stats like the blockID store did.

### Does this PR introduce _any_ user-facing change?

`spark.rss.client.integrityValidation.serverManagementEnabled=false`

### How was this patch tested?

Internal job tests.